On Windows, add base directories like \\?\c:\
The path._makeLong function will return paths like these on Windows:

    \\?\c:\foo\bar

The fs.realpath function walks through Windows paths treating '\\?\c:\' as the base directory, so this is what we store.  The process.cwd and os.tmpdir functions may not return the same case for base directory names, so we normalize to lowercase.