Add function constructor for files.
This lets files be populated async, e.g.: creating a mock overlay of a
live file system. The function may return `fs.readFile(...)`, pull a
buffer from S3 or GitHub, or arbitrary other IO actions to return a
string or Buffer.

This is not intended to allow populating directories async, it's assumed
that crawling directories is cheap relative to reading files.