Update fs.stat(), fs.lstat(), and fs.fstat() for Node 7.7+ (fixes #197)
Node 7.7 changed the behavior of the `binding.{stat,lstat,fstat}` functions (see https://github.com/nodejs/node/pull/11522). This updates the binding functions used in mock-fs to match the new Node binding behavior, while still maintaining compatibility with old Node versions. The new behavior is detected when the second argument to `binding.{stat,lstat,fstat}` is a `Float64Array`, which would be an invalid argument for previous versions of the binding.