fix: throw ENOTDIR when trying to open an incorrect path (nested under existing file)
If you have a file: /mock-dir/two.txt and try to open /mock-dir/two.txt/bogus-path, you would get `item.getItem is not a function` due to `item` instance being a File. This adds a fix to this by checking the instance of resolved `item` and throwing accordingly. This matches node's native behavior in such a case.