fix FileHandle write on node 20
Both `fs.promises.open().then(fh => fh.write()` and `fs.openSync().write()`
call write without a callback, but fs.promises must return a Promise,
while openSync must not.

I added _isPromise to the descriptor so we can return the right type
for the handle open was called with.

This doesn't feel right, but I couldn't figure out a different way to
call it.