fix: encoding utf8 strings in node 20
The writeFileUtf8 binding used by node 20 was not correctly encoding
strings as utf8, and recording the file length as
String.prototype.length, which returns number of utf-16 code units.

If you wrote a string containing code points with a different number of
utf-8 code units than utf-16 code units, then read the file back the
returned string would be truncated by the difference in the number of
code units.